Added the Hasher and LoginUtility classes
With this most interesting commit I intend to add long needed features to the project. Via the Husher class you can, passing a string as a parameter, get back a nicely encrypted string you can use to impress all the other kinds in the block. With the LoginUtility class it was my goal to check if a user was registered in the chat's database and then return to the requester a positive or negative acknowledgment. Hope it works 'cause sure as he*l i didn't test it