Added RegisterUtility class
The class is comprised of two methods in order to divide better the responsibilities. The isRegistered method checks if the username passed is already present in the database. If not it returns a positive response. The register method does all the good stuff and eventually adds a user to the database with its username and hashed password. All this sh*t is not tested btw.